Add the subtract multiply divide functionality
I had no access to the git today. I was making changes in college so
some parts of the code just pop out of nowhere.

Also at this point i think my code has several clutches, hard to read
and cluttered with parts that can be turn into a function.

And yeah, though it is not my first attempt on making a calculator
the first one i have made with help of a tutorial on youtube,
but this time i was making it from scratch by my self, please don't
judge it too harsh, just say what i could've made better